Add LivView to PartIdentification
