+++ factory-ai-vision/EdgeSolution/modules/WebModule/ui/src/components/CameraConfigure/index.tsx
@@ -1,4 +1,3 @@
-import { CameraConfigureInfoContainer } from './CameraConfigureInfo';
 import { CreateCameraConfigButton } from './CreateCameraConfig';
 
-export { CameraConfigureInfoContainer, CreateCameraConfigButton };
+export { CreateCameraConfigButton };
