+++ factory-ai-vision/EdgeSolution/modules/WebModule/ui/src/components/LiveViewDashboard/index.tsx
@@ -1,0 +1,3 @@
+import { LiveViewDashboard } from './LiveViewDashboard';
+
+export { LiveViewDashboard };
