+++ factory-ai-vision/EdgeSolution/modules/WebModule/ui/src/pages/PartIdentification.tsx
@@ -1,16 +1,16 @@
 import React from 'react';
 import { Flex } from '@fluentui/react-northstar';
 import { ProjectConfig } from '../components/ProjectConfigPanel/ProjectConfig';
-import { LiveViewContainer } from '../components/LiveViewContainer';
+import { LiveViewDashboard } from '../components/LiveViewDashboard';
 
 export const PartIdentification: React.FC = () => {
   return (
-    <Flex>
+    <Flex styles={{ height: '100%' }}>
       <div style={{ flexGrow: 2 }}>
         <ProjectConfig />
       </div>
       <div style={{ flexGrow: 3 }}>
-        <LiveViewContainer showVideo={true} initialAOIData={{ useAOI: false, AOIs: [] }} cameraId={0} />
+        <LiveViewDashboard projectId={0} />
       </div>
     </Flex>
   );
